optimize discussion data fetching
